+++ tools/PSc8y/Public-manual/Test-ClientPassphrase.ps1
@@ -37,7 +37,11 @@ Set the passphrase if it is not already set
             return
         }
 
-        $encryptionInfo = $passphraseCheck | ConvertFrom-Json
+        $JSONArgs = @{}
+        if ($PSVersionTable.PSVersion.Major -gt 5) {
+            $JSONArgs.Depth = 100
+        }
+        $encryptionInfo = $passphraseCheck | ConvertFrom-Json @JSONArgs
 
         # Save passphrase to env variable
         $env:C8Y_PASSWORD = $encryptionInfo.C8Y_PASSWORD